LG-11285: Add helper function to clean up observable properties (#10809)
* Extract defineObservableProperty into its own file

* Add removeObservableProperty

* Use removeObservableProperty in acuant-capture-canvas

* Rename to stopObservingProperty

* Fix issues due to merge, lint issues

* Another lint fix

* Remove unused

* Typescript convert so we don't have to add any exceptions
